refactor: 기존 ResetPassword 로직 활용
