+++ app/src/main/java/com/android/quizcafe/core/data/remote/datasource/UserRemoteDataSource.kt
@@ -1,7 +1,7 @@
 package com.android.quizcafe.core.data.remote.datasource
 
+import com.android.quizcafe.core.data.model.auth.request.ResetPasswordRequestDto
 import com.android.quizcafe.core.data.model.quizbook.response.QuizBookResponseDto
-import com.android.quizcafe.core.data.model.user.request.ResetPasswordRequestDto
 import com.android.quizcafe.core.data.model.user.response.UserInfoResponseDto
 import com.android.quizcafe.core.data.remote.service.UserService
 import com.android.quizcafe.core.network.model.ApiResponse
@@ -15,13 +15,10 @@ class UserRemoteDataSource @Inject constructor(
     suspend fun getUserInfo(): NetworkResult<ApiResponse<UserInfoResponseDto>> =
         userService.getUserInfo()
 
-    suspend fun deleteUser(): NetworkResult<Unit> = userService.deleteUser()
+    suspend fun deleteUser(): NetworkResult<ApiResponse<Unit>> = userService.deleteUser()
 
-    suspend fun updateUserNickName(nickName: String): NetworkResult<Unit> =
+    suspend fun updateUserNickName(nickName: String): NetworkResult<ApiResponse<Unit>> =
         userService.updateUserNickName(nickName = nickName)
-
-    suspend fun resetPassword(request: ResetPasswordRequestDto): NetworkResult<Unit> =
-        userService.resetPassword(request = request)
 
     suspend fun getMyQuizBooks(): NetworkResult<ApiResponse<List<QuizBookResponseDto>>> =
         userService.getMyQuizBooks()
