+++ app/src/main/java/com/android/quizcafe/core/domain/model/auth/request/ResetPasswordRequest.kt
@@ -1,5 +1,6 @@
 package com.android.quizcafe.core.domain.model.auth.request
 
 data class ResetPasswordRequest(
-    val email: String
+    val oldPassword: String,
+    val newPassword: String
 )
