--- app/src/main/java/com/android/quizcafe/core/domain/model/user/requst/ResetPasswordRequest.kt
@@ -1,6 +1,0 @@
-package com.android.quizcafe.core.domain.model.user.requst
-
-data class ResetPasswordRequest(
-    val oldPassword: String,
-    val newPassword: String
-)
