+++ app/src/main/java/com/android/quizcafe/core/domain/repository/UserRepository.kt
@@ -2,7 +2,6 @@ package com.android.quizcafe.core.domain.repository
 
 import com.android.quizcafe.core.domain.model.Resource
 import com.android.quizcafe.core.domain.model.quizbook.response.QuizBook
-import com.android.quizcafe.core.domain.model.user.requst.ResetPasswordRequest
 import com.android.quizcafe.core.domain.model.user.response.UserInfo
 import kotlinx.coroutines.flow.Flow
 
@@ -10,7 +9,6 @@ interface UserRepository {
 
     fun getUserInfo(): Flow<Resource<UserInfo>>
     fun updateUserNickName(nickName: String): Flow<Resource<Unit>>
-    fun resetPassword(request: ResetPasswordRequest): Flow<Resource<Unit>>
     fun deleteUser(): Flow<Resource<Unit>>
     fun getMyQuizBooks(): Flow<Resource<List<QuizBook>>>
 }
