+++ app/src/main/java/com/android/quizcafe/core/domain/usecase/auth/ResetPasswordUseCase.kt
@@ -9,6 +9,6 @@ import javax.inject.Inject
 class ResetPasswordUseCase @Inject constructor(
     private val authRepository: AuthRepository
 ) {
-    operator fun invoke(request: ResetPasswordRequest): Flow<Resource<Unit>> =
-        authRepository.resetPassword(request)
+    operator fun invoke(oldPassword: String, newPassword: String): Flow<Resource<Unit>> =
+        authRepository.resetPassword(ResetPasswordRequest(oldPassword = oldPassword, newPassword = newPassword))
 }
